+++ ocaz-face-detector-deepface/src/sandbox/detect_retina_face.py
@@ -2,13 +2,12 @@ import re
 
 import cv2
 import numpy as np
-from retinaface import RetinaFace
-from retinaface.commons import postprocess
+from ocaz_face_detector_deepface.face_detector import RetinaFaceDetector
 
 
 # https://github.com/serengil/deepface/blob/ce4e4f664b66c05e682de8c0913798da0420dae1/deepface/commons/functions.py#L119
 def resize(image: np.ndarray, width: int, height: int) -> np.ndarray:
-    image_height, image_width, _ = face_image.shape
+    image_height, image_width, _ = image.shape
     factor = min(height / image_height, width / image_width)
     resized_image = cv2.resize(
         image,
@@ -47,52 +46,7 @@ def resize_with_pad(image: np.ndarray, width: int, height: int) -> np.ndarray:
     return pad(resize(image, width, height), width, height)
 
 
-class RetinaFaceDetector:
-    def __init__(self):
-        RetinaFace.build_model()
-
-    def detect(self, image, threshold=0.5, allow_upscaling=True):
-        def array_to_dict(x, y):
-            return {"x": x, "y": y}
-
-        faces = RetinaFace.detect_faces(image, threshold=threshold, allow_upscaling=allow_upscaling)
-
-        results = []
-        for _, face in faces.items():
-            x1, y1, x2, y2 = face["facial_area"]
-            face_image = image[y1:y2, x1:x2]
-
-            landmarks = face["landmarks"]
-            left_eye = landmarks["left_eye"]
-            right_eye = landmarks["right_eye"]
-            nose = landmarks["nose"]
-            mouth_right = landmarks["mouth_right"]
-            mouth_left = landmarks["mouth_left"]
-
-            aligned_image = postprocess.alignment_procedure(face_image, right_eye, left_eye, nose)
-
-            results.append(
-                {
-                    "score": face["score"],
-                    "boundingBox": {"x1": x1, "y1": y1, "x2": x2, "y2": y2},
-                    "landmarks": {
-                        "leftEye": array_to_dict(*left_eye),
-                        "rightEye": array_to_dict(*right_eye),
-                        "nose": array_to_dict(*nose),
-                        "mouthRight": array_to_dict(*mouth_right),
-                        "mouthLeft": array_to_dict(*mouth_left),
-                    },
-                    "alignedImage": aligned_image,
-                }
-            )
-
-        return results
-
-
 image = cv2.imread("image1.jpg")
-
-# resp = RetinaFace.detect_faces(image, threshold=0.5)
-# print(resp)
 
 face_detector = RetinaFaceDetector()
 faces = face_detector.detect(image)
@@ -100,13 +54,8 @@ for i, face in enumerate(faces):
     aligned_image = face["alignedImage"]
     cv2.imwrite(f"face_{i}.jpg", aligned_image)
 
-    # print(faces)
-
-# resp = RetinaFace.extract_faces(image, threshold=0.5)
-# print(resp[0].shape)
-# print(resp[0].dtype)
-# cv2.imwrite("face1.jpg", resp[0][:, :, ::-1])
-
+    del face["alignedImage"]
+    print(face)
 
 # face_image = resp[0]
